+++ tests/PhpSpreadsheetTests/Helper/HtmlTest.php
@@ -48,4 +48,26 @@ class HtmlTest extends TestCase
 
         self::assertSame($expected, $actual->getPlainText());
     }
+
+    public function testSTag(): void
+    {
+        $html = new Html();
+        $input = 'Hello <s>test</s>world';
+        $richText = $html->toRichTextObject($input);
+        $elements = $richText->getRichTextElements();
+
+        self::assertSame(count($elements), 3);
+
+        self::assertSame($elements[0]->getText(), 'Hello ');
+        self::assertNotNull($elements[0]->getFont());
+        self::assertFalse($elements[0]->getFont()->getStrikethrough());
+
+        self::assertSame($elements[1]->getText(), 'test');
+        self::assertNotNull($elements[1]->getFont());
+        self::assertTrue($elements[1]->getFont()->getStrikethrough());
+
+        self::assertSame($elements[2]->getText(), 'world');
+        self::assertNotNull($elements[2]->getFont());
+        self::assertFalse($elements[2]->getFont()->getStrikethrough());
+    }
 }